requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.1.1
absl-py=2.2.2=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
astunparse=1.6.3=pypi_0
attrs=25.3.0=pypi_0
bcrypt=4.3.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.2.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cryptography=44.0.2=pypi_0
customtkinter=5.2.2=pypi_0
cycler=0.12.1=pypi_0
darkdetect=0.8.0=pypi_0
dash=3.0.4=pypi_0
dash-bootstrap-components=2.0.2=pypi_0
debugpy=1.8.11=py312h5da7b33_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.7.1=h8ddb27b_0
flask=3.0.3=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.57.0=pypi_0
gast=0.6.0=pypi_0
google-pasta=0.2.0=pypi_0
graphviz=0.20.3=pypi_0
grpcio=1.71.0=pypi_0
h5py=3.13.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh4bbf305_0
ipython=9.1.0=pyhca29cf9_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
itsdangerous=2.2.0=pypi_0
jax=0.6.0=pypi_0
jaxlib=0.6.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=py312haa95532_0
keras=3.9.2=pypi_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=hdf4eb48_0
libclang=18.1.1=pypi_0
libffi=3.4.4=hd77b12b_1
libsodium=1.0.20=hc70643c_0
markdown=3.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mediapipe=0.10.21=pypi_0
ml-dtypes=0.5.1=pypi_0
namex=0.0.9=pypi_0
narwhals=1.39.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
numpy=1.26.4=pypi_0
opencv-contrib-python=4.11.0.86=pypi_0
openssl=3.5.0=ha4e3fda_0
opt-einsum=3.4.0=pypi_0
optree=0.15.0=pypi_0
packaging=25.0=pyh29332c3_1
pandas=2.2.3=pypi_0
paramiko=3.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=pypi_0
pip=25.0=py312haa95532_0
platformdirs=4.3.7=pyh29332c3_0
plotly=6.0.1=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
protobuf=4.25.6=pypi_0
psutil=5.9.0=py312h827c3e9_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pypi_0
pydot=4.0.0=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pynacl=1.5.0=pypi_0
pyparsing=3.2.3=pypi_0
python=3.12.6=h14ffc60_1
python-dateutil=2.9.0.post0=pyhff2d567_1
pytz=2025.2=pypi_0
pywin32=308=py312h5da7b33_0
pyzmq=26.2.0=py312h5da7b33_0
requests=2.32.3=pypi_0
retrying=1.3.4=pypi_0
rich=14.0.0=pypi_0
scipy=1.15.2=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=75.8.0=py312haa95532_0
six=1.17.0=pyhd8ed1ab_0
sounddevice=0.5.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.3=pyhd8ed1ab_1
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.16.1=pypi_0
tensorflow-intel=2.16.1=pypi_0
termcolor=3.0.1=pypi_0
tk=8.6.14=h0416ee5_0
tornado=6.4.2=py312h827c3e9_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.4.0=pypi_0
vc=14.42=haa95532_5
vc14_runtime=14.42.34438=hfd919c2_26
vs2015_runtime=14.42.34438=h7142326_26
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.0.6=pypi_0
wheel=0.45.1=py312haa95532_0
wrapt=1.17.2=pypi_0
xz=5.6.4=h4754444_1
zeromq=4.3.5=ha9f60a1_7
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h8cc25b3_1
